requirements.txt
python-dotenv==0.15.0
PyYAML==5.4.1
requests==2.25.1
